requirements.txt
absl-py==2.3.1
audioread==3.1.0
cachetools==4.2.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dataclassy==1.0.1
decorator==5.2.1
Distance==0.1.3
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.10.0
g2p-en==2.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.76.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==4.1.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
kneed==0.8.5
librosa==0.7.2
llvmlite==0.43.0
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
Montreal_Forced_Aligner==3.3.8
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.2
numba==0.60.0
numpy==2.0.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==11.3.0
praatio==6.2.0
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
pyopenjtalk==0.4.1
pyparsing==3.2.5
pypinyin==0.39.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyworld-prebuilt==0.3.5.post2
PyYAML==6.0.1
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
resampy==0.4.3
rich==14.2.0
rich-click==1.9.4
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
SoundFile==0.10.3.post1
SQLAlchemy==2.0.44
sympy==1.14.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tgt==1.4.4
threadpoolctl==3.6.0
torch==2.8.0
tqdm==4.46.1
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.1.1
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
